Minor - unsuccessful debugging

+++ src/main.py
@@ -91,7 +91,7 @@ import plot_functions as pf
 
 
 def prepare_structure(path=SIC_ROOT,
-                      skip=[SIC_ORIG, SIC_SCRIPTS, "orig", "orig1", "orig2", "orig3", "orig4", "orig5"],
+                      skip=[SIC_ORIG, SIC_SCRIPTS, "orig", "orig1", "orig2", "orig3", "orig4", "orig5", "orig6"],
                       create_dirs=[SIC_PROCESSED, SIC_RESULTS, SIC_LINKS],
                       check_for=[join(SIC_ROOT, SIC_SCRIPTS, FIJI_STANDARD_SCRIPT),
                         join(SIC_ROOT, SIC_ORIG)]
@@ -103,7 +103,7 @@ def prepare_structure(path=SIC_ROOT,
         l = listdir(path)
         for i in sorted(l):
             # removing everything which is not a SIC_ORIG or SIC_SCRIPTS
-            if i not in skip:
+            if i not in skip and i[:3]!='orig':
                 rmtree(join(path, i))
                 print "Removing:", join(path, i)
             else:
@@ -575,23 +575,34 @@ def aggregate_and_track_spots(spots, niba2dic):
             return False
     
     def isClosestSuccessor(spot1, spot2, spots):
-        if len(spots) <= 2 and slice_distance(spot1, spot2) == 1: 
-            return True
-        if dist2((spot1[2], spot1[3]), (spot2[2], spot2[3])) == min([dist2((spot1[2], spot1[3]), (spot[2], spot[3])) for spot in spots if spot!=spot1])\
-        and slice_distance(spot1, spot2) == 1:
-            return True
+        # is spot2 the closest (minimum distance) successor (on the next slice) of spot 1
+        if slice_distance(spot1, spot2) == 1:
+            if dist2((spot1[2], spot1[3]), (spot2[2], spot2[3])) == min([dist2((spot1[2], spot1[3]), (spot[2], spot[3])) for spot in spots\
+                                                                         if spot!=spot1 and slice_distance(spot1, spot) == 1]):
+                return True
         else:
             return False
     
     def isClosestPredecessor(spot1, spot2, spots):
-        if len(spots) <= 2 and slice_distance(spot1, spot2) == -1: 
-            return True
-        if dist2((spot1[2], spot1[3]), (spot2[2], spot2[3])) == min([dist2((spot1[2], spot1[3]), (spot[2], spot[3])) for spot in spots if spot!=spot1])\
-        and slice_distance(spot1, spot2) == -1:
-            return True
+        # is spot2 the closest (minimum distance) predecessor (on the previous slice) of spot 1
+        if slice_distance(spot1, spot2) == -1:
+            if dist2((spot1[2], spot1[3]), (spot2[2], spot2[3])) == min([dist2((spot1[2], spot1[3]), (spot[2], spot[3])) for spot in spots\
+                                                                         if spot!=spot1 and slice_distance(spot1, spot) == -1]):
+                return True
         else:
             return False
 
+    def onTrajectoryNew(spot1, spot2, spots):
+        if spot1 == spot2:
+            return 'Y' # yes
+        elif dist2((spot1[2], spot1[3]), (spot2[2], spot2[3])) >= slice_distance(spot1, spot2)*slice_distance(spot1, spot2)*CRIT_DIST2_FROM_PREV:
+            return 'N' # no
+        elif isWithinDistance(spot1, spot2) and (isClosestSuccessor(spot1, spot2, spots) or isClosestSuccessor(spot2, spot1, spots))\
+        and (isClosestPredecessor(spot1, spot2, spots) or isClosestPredecessor(spot2, spot1, spots)):
+            return 'Y' # yes
+        else:
+            return 'M' # maybe
+    
     def onTrajectory(spot1, spot2, spots):
         '''
         Functional programming implementation of spot trajectory:
@@ -613,11 +624,10 @@ def aggregate_and_track_spots(spots, niba2dic):
                     #print "\t\tRecursing on spot:", spot3
                     if len(spots)<=2:
                         return False 
-                    spotsminusspot1 = deepcopy(spots) 
-                    spotsminusspot2 = deepcopy(spots) 
+                    spots1 = deepcopy(spots) 
+                    spots2 = deepcopy(spots) 
                     if spot3!=spot1 and spot3!=spot2:
-                        if onTrajectory(spot3, spot1, spotsminusspot2) and onTrajectory(spot3, spot2, spotsminusspot1):
-                            return True
+                        return (onTrajectory(spot3, spot1, spots2) and onTrajectory(spot3, spot2, spots1)) or (onTrajectory(spot1, spot3, spots2) and onTrajectory(spot2, spot1, spots1))
             return False
     
     stacks = sorted(list(set([niba2dic[spot[-1]] for spot in spots]))) # spot[-1] is the old file ID (can be -max.tif or belonging to slice)
@@ -625,7 +635,7 @@ def aggregate_and_track_spots(spots, niba2dic):
     for stack in stacks:
         spotted_cells = sorted(list(set([spot[1] for spot in spots if niba2dic[spot[-1]]==stack and spot[-1].find(CELLID_FP_TOKEN)!=-1])))
         # condition 1 selects only cells in the current stack
-        # condition 2 makes sure that only spots contained in the max projections are considered real (i.e. not those in the slices)
+        # condition 2 makes sure that only spots contained in the max projections are considered real (i.e. not those that are only in the slices)
 
         max_projection_spots = [spot for spot in spots if niba2dic[spot[-1]]==stack and spot[-1].find(CELLID_FP_TOKEN)!=-1]
         # this is the 'master spot list' containing only spots that are visible on the max projection
@@ -639,6 +649,31 @@ def aggregate_and_track_spots(spots, niba2dic):
             print "The following spots were found:"
             for spot in local_spots:
                 print spot
+            
+            '''
+            print "---------------------------------------------"
+            num_spots = len(local_spots)                
+            
+            # Initialisation
+            trajectory_matrix = [['M' for k in range(num_spots)] for l in range(num_spots)]
+            
+            for k in range(num_spots):
+                print trajectory_matrix[k]
+            print "---------------------------------------------"
+            
+            for k, spotk in enumerate(local_spots):
+                for l, spotl in enumerate(local_spots):
+                    trajectory_matrix[k][l] = onTrajectory(spotk, spotl, local_spots)
+                    trajectory_matrix[l][k] = trajectory_matrix[k][l]
+
+            for k in range(num_spots):
+                print trajectory_matrix[k]
+            print "---------------------------------------------"
+
+            trajectory_matrix_string = "".join(trajectory_matrix[i][j] for i in range(num_spots) for j in range(num_spots))
+            print trajectory_matrix_string.find('M') != -1 # are positions still undecided?
+            '''
+
             '''
             for spot1 in local_spots:
                 for spot2 in local_spots:
@@ -647,19 +682,30 @@ def aggregate_and_track_spots(spots, niba2dic):
                     print spot2
                     print "isWithinDistance(spot1, spot2):", isWithinDistance(spot1, spot2)
                     print "isClosestSuccessor(spot1, spot2, spots) or isClosestSuccessor(spot2, spot1, spots):", isClosestSuccessor(spot1, spot2, spots) or isClosestSuccessor(spot2, spot1, spots)
+                    print "isClosestPredecessor(spot1, spot2, spots) or isClosestPredecessor(spot2, spot1, spots):", isClosestPredecessor(spot1, spot2, spots) or isClosestPredecessor(spot2, spot1, spots)
                     print "onTrajectory(spot1, spot2, local_spots):", onTrajectory(spot1, spot2, local_spots)
                     raw_input('Press Enter...')
             '''
+            
+
             for spot1 in local_spots:
                 print "---------------------------------------------"
                 print spot1
                 for spot in local_spots:
-                    if onTrajectory(spot, spot1, local_spots):
+                    if onTrajectory(spot, spot1, local_spots) or onTrajectory(spot1, spot, local_spots):
                         print "\t", spot
+            print "---------------------------------------------"
+
+
+            print "---------------------------------------------"
+            spot1 = ['GFP_P020005_T0', '8', 569.549866272399, 597.113272100496, 214.0, 534373.0, 208451.0, 1523.0, 252.042, 270.88895334572567, 0.0, 'Stack_02_w2NIBA-0005.tif']
+            spot2 = ['GFP_P020006_T0', '8', 569.365923776806, 596.730425964634, 166.0, 425722.0, 162612.0, 1585.0, 257.2311, 211.31966016692243, 0.0, 'Stack_02_w2NIBA-0006.tif']
+            print isWithinDistance(spot1, spot2)
+            print (isClosestSuccessor(spot1, spot2, local_spots) or isClosestSuccessor(spot2, spot1, local_spots))
+            print (isClosestPredecessor(spot1, spot2, local_spots) or isClosestPredecessor(spot2, spot1, local_spots))
+            print onTrajectory(spot1, spot2, local_spots)  
+
             
-            print "---------------------------------------------"
-            print "---------------------------------------------"
-            print onTrajectory(['GFP_P020005_T0', '8', 569.549866272399, 597.113272100496, 214.0, 534373.0, 208451.0, 1523.0, 252.042, 270.88895334572567, 0.0, 'Stack_02_w2NIBA-0005.tif'], ['GFP_P020003_T0', '8', 570.60583496376, 597.138818032629, 184.0, 409334.0, 135450.0, 1488.5, 276.5049, 176.0217448257794, 0.0, 'Stack_02_w2NIBA-0003.tif'], local_spots)
             '''
             for spot1ID, spot1 in enumerate(local_spots):
                 sliceID[spot1ID] = int(spot1[0].split("_")[1][-4:])
